requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-http-client==3.3.7
sendgrid==6.12.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
